feat: some new tests implemented.

+++ src/main/java/it/polito/ezshop/data/SQLiteDB.java
@@ -1,5 +1,6 @@
 package it.polito.ezshop.data;
 
+import java.lang.reflect.Array;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
@@ -8,6 +9,7 @@ import java.sql.Statement;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
@@ -18,7 +20,19 @@ public class SQLiteDB {
     static final int defaultID = -1;
     static final int defaultValue = 0;
     private static final int INTEGER = 4; // see https://docs.oracle.com/javase/8/docs/api/constant-values.html#java.sql.Types.INTEGER
-    Connection dbConnection = null;
+
+    // Define tables names
+    private static final String tBalanceOperations = "BalanceOperations";
+    private static final String tCards = "Cards";
+    private static final String tCustomers = "Customers";
+    private static final String tOrders = "Orders";
+    private static final String tProductsPerSale = "ProductsPerSale";
+    private static final String tProductTypes = "ProductTypes";
+    private static final String tTransactions = "Transactions";
+    private static final String[] tables =  new String [] {tBalanceOperations, tCards, tCustomers, tOrders, tProductsPerSale,
+                                                            tProductsPerSale, tProductTypes, tTransactions};
+
+    private Connection dbConnection = null;
 
     /**
      ** Connect to the DB
@@ -96,24 +110,36 @@ public class SQLiteDB {
         if (this.dbConnection == null)
             return false;
 
-        this.clearTable("BalanceOperations");
-        this.clearTable("Orders");
-        this.clearTable("ProductsPerSale");
-        this.clearTable("ProductTypes");
-        this.clearTable("Transactions");
+        boolean cleared = this.clearTable(tBalanceOperations);
+        cleared &= this.clearTable(tOrders);
+        cleared &= this.clearTable(tProductsPerSale);
+        cleared &= this.clearTable(tProductTypes);
+        cleared &= this.clearTable(tTransactions);
 
-        return true;
+        return cleared;
   }
 
-  private void clearTable(String tableName) {
-      String sql =  "DELETE FROM " + tableName + " ;";
+  private boolean clearTable(String tableName) {
 
-      try{
-          Statement stmt = this.dbConnection.createStatement();
-          stmt.execute(sql);
-      } catch (SQLException e) {
-          System.out.println(e.getMessage());
-      }
+        // Validate tableName, in order to prevent SQL injections
+        boolean isValidTable = false;
+        for (String tName : SQLiteDB.tables)
+            isValidTable |= tableName.equals(tName);
+
+        if (!isValidTable)
+            return false;
+
+        String sql =  "DELETE FROM " + tableName + " ;";
+
+        try{
+            Statement stmt = this.dbConnection.createStatement();
+            stmt.execute(sql);
+            return true;
+        } catch (SQLException e) {
+            System.out.println(e.getMessage());
+        }
+
+        return false;
   }
 
     /**
@@ -986,7 +1012,8 @@ public class SQLiteDB {
         HashMap<Integer, EZSaleTransaction> saleTransactions = new HashMap<>();
         String sql = "SELECT id, discountRate, price, status \n"
                    + "FROM Transactions \n"
-                   + "WHERE status <> \"OPENED\";";
+                   + "WHERE saleID IS NULL \n"
+                   + "  AND status <> \"OPENED\";";
 
         try {
             Statement stmt  = this.dbConnection.createStatement();
